Create README.md
We are given a set of dog and cat images. Using these images, you need to train a deep learning image classification model, with descent accuracy of at least 90%. The task is to build a model to predict the category of an animal: dog or cat if an image is passed to the model.